+++ src/main/java/SN/BANK/users/dto/LoginDto.java
@@ -1,4 +1,4 @@
-package SN.BANK.dto;
+package SN.BANK.users.dto;
 
 import jakarta.validation.constraints.NotBlank;
 import lombok.Builder;
@@ -7,19 +7,15 @@ import lombok.NoArgsConstructor;
 
 @Getter
 @NoArgsConstructor
-public class UsersRequestDto {
-    @NotBlank
-    private String name;
+public class LoginDto {
     @NotBlank
     private String loginId;
     @NotBlank
     private String password;
 
     @Builder
-    public UsersRequestDto(String name, String loginId, String password){
-        this.name = name;
-        this.loginId= loginId;
+    public LoginDto(String loginId, String password){
+        this.loginId = loginId;
         this.password = password;
     }
-
 }
