+++ src/main/java/SN/BANK/users/repository/UsersRepository.java
@@ -1,6 +1,6 @@
-package SN.BANK.repository;
+package SN.BANK.users.repository;
 
-import SN.BANK.domain.Users;
+import SN.BANK.users.entity.Users;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
